电商数据采集测试

+++ product/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Product(models.Model):
     title = models.CharField(max_length=120)
     desc = models.TextField()
-    price = models.PositiveIntegerField()
+    price = models.DecimalField(decimal_places=2, max_digits=10)
     update_time = models.DateTimeField(auto_now=True)
     create_time = models.DateTimeField(auto_now_add=True)
     is_active = models.BooleanField(default=True)
